Rename base service file
