+++ api/src/services/dwc-service.ts
@@ -6,7 +6,7 @@ import { Queries } from '../queries';
 import { DWCArchive } from '../utils/media/dwc/dwc-archive-file';
 import { ArchiveFile } from '../utils/media/media-file';
 import { parseUnknownMedia } from '../utils/media/media-utils';
-import { DBService } from './service';
+import { DBService } from './db-service';
 
 export class DarwinCoreService extends DBService {
   /**
